Rename NewFile1.cpp to Assignment no.2
Multilevel Inheritance in C++

This C++ program demonstrates the concept of multilevel inheritance, where a class inherits from another class, which in turn inherits from another. In this example, we have three classes: sportBike, Twowheeler, and Bike.

sportBike: The base class that contains a method to display a message about sportbikes.

Twowheeler: A derived class that inherits from sportBike and adds a method to display information about two-wheelers.

Bike: A class that further inherits from Twowheeler and contains a method to display information about bikes.


In the main() function, we create an object of the Bike class, which allows access to methods from all the inherited classes.